bill issue added

+++ op-bo/onlinepayment-bo/src/main/java/af/asr/opbo/opbo/model/FeeModel.java
@@ -21,7 +21,7 @@ public class FeeModel extends BaseEntity {
     private String name;
     private  boolean isItemBased;
     private String type; //percentage or exact amount
-    private Double percentage;
+    private BigDecimal percentage;
     private BigDecimal amount;
 
 }
